Advancing Nutrition Science Through GeminiAI/requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
python-dotenv
pillow

